// src/lib/types.ts
// Services type definition
export interface Services {
    electricalServices: {
      title: string;
      type: string;
    };
    plumbingServices: {
      title: string;
      type: string;
    };
    carpentryServices: {
      title: string;
      type: string;
    };
    paintingServices: {
      title: string;
      type: string;
    };
    handymanServices: {
      title: string;
      type: string;
    };
    applianceRepairServices: {
      title: string;
      type: string;
    };
    roofingServices: {
      title: string;
      type: string;
    };
    cleaningServices: {
      title: string;
      type: string;
    };
  }